Butterfly and SISL Renaming
