+++ pettingzoo/butterfly/cooperative_pong_v5.py
@@ -1,1 +1,1 @@
-from .cooperative_pong.cooperative_pong import ManualPolicy, env, parallel_env, raw_env
+from .cooperative_pong.cooperative_pong_env import ManualPolicy, env, parallel_env, raw_env
