+++ pettingzoo/butterfly/knights_archers_zombies/knights_archers_zombies_env.py
