+++ pettingzoo/butterfly/knights_archers_zombies_v10.py
@@ -1,4 +1,4 @@
-from .knights_archers_zombies.knights_archers_zombies import (
+from .knights_archers_zombies.knights_archers_zombies_env import (
     ManualPolicy,
     env,
     parallel_env,
