+++ pettingzoo/butterfly/pistonball_v6.py
@@ -1,1 +1,1 @@
-from .pistonball.pistonball import ManualPolicy, env, parallel_env, raw_env
+from .pistonball.pistonball_env import ManualPolicy, env, parallel_env, raw_env
