+++ pettingzoo/butterfly/prospector_v4.py
@@ -1,1 +1,1 @@
-from .prospector.prospector import env, manual_control, parallel_env, raw_env
+from .prospector.prospector_env import env, manual_control, parallel_env, raw_env
